ofproto-dpif: Remove unneeded bundle_remove() from port_del().
The code in this function to call bundle_remove() dates back to before
netdevs had a reference count.  At that time, when the caller closed the
ofport's netdev, any further reference to that netdev from the bond would
access freed memory.  Now that netdevs are ref-counted, it's no longer a
problem, so removed the call.

Signed-off-by: Ben Pfaff <blp@nicira.com>