[RFC] Add support for capturing all OpenFlow traffic to a .pcap file.
Occasionally I get asked for the ability to capture all OpenFlow traffic
in a raw format, to allow for analysis and debugging.  OVS has a few
existing tools for this but none of them really fit the bill.  For example,
turning up the log level for the "vconn" module will dump all traffic but
only in a pre-parsed format, and "ovs-ofctl snoop" can only print a single
connection at a time (also pre-parsed).  One can instead run tcpdump to
capture all traffic, but then you run into the issue that you need the SSL
keys to decrypt it, which is inconvenient.

This commit proposes another approach: write all OpenFlow traffic to a
.pcap file in binary format.  The really odd part of this approach is that
OVS doesn't actually have the headers that should go into the .pcap file
(because OVS uses the host network stack) so we have to synthesize
plausible ones.  While that's easy enough (it's about 100 lines of code in
this commit), it's weird enough that it seems worth doing an RFC patch for
it.

This commit is pretty preliminary in that it doesn't have a proper approach
for configuring and enabling the feature.  Instead, it just writes
everything that goes over a TCP stream socket to a file in the current
directory called capture.pcap.

Feedback appreciated!