tests: Automatically initialize OVS_*DIR vars when tests begin.
A lot of tests need to initialize the OVS_RUNDIR, OVS_LOGDIR, etc.
variables to point to the directory in which the tests run.  Until now,
each of them has had to do this individually, which is redundant.  This
commit starts to do this automatically.

The ovs_setenv function introduced in this commit is more generally
applicable than this commit takes advantage of.  Future commits will use
its additional power.

Signed-off-by: Ben Pfaff <blp@nicira.com>