ovn: Implement basic end-to-end full mesh test.
This is a really basic test of the OVN features.  It verifies that basic
L2 connectivity works as expected over a 3-hypervisor setup with 3 VMs
per hypervisor and all 9 VMs on a single logical switch.

I wanted to test stateless ACLs also but there's no simple way to set them
up with ovn-nbctl.  I guess that points toward a future direction!

The infrastructure added by this patch, which is based on similar code
from ovs-sim, should be useful as a basis for later and more advanced
OVN end-to-end tests.

Signed-off-by: Ben Pfaff <blp@nicira.com>