Implement "closures".
One purpose of OpenFlow packet-in messages is to allow a controller to
interpose on the path of a packet through the flow tables.  If, for
example, the controller needs to modify a packet in some way that the
switch doesn't directly support, the controller should be able to
program the switch to send it the packet, then modify the packet and
send it back to the switch to continue through the flow table.

That's the theory.  In practice, this doesn't work with any but the
simplest flow tables.  Packet-in messages simply don't include enough
context to allow the flow table traversal to continue.  For example:

    * Via "resubmit" actions, an Open vSwitch packet can have an
      effective "call stack", but a packet-in can't describe it, and
      so it would be lost.

    * Via "patch ports", an Open vSwitch packet can traverse multiple
      OpenFlow logical switches.  A packet-in can't describe or resume
      this context.

    * A packet-in can't preserve the stack used by NXAST_PUSH and
      NXAST_POP actions.

    * A packet-in can't preserve the OpenFlow 1.1+ action set.

    * A packet-in can't preserve the state of Open vSwitch mirroring
      or connection tracking.

This commit introduces a solution called "closures".  A closure is the
state of a packet's traversal through OpenFlow flow tables.  A new
NXAST_PAUSE action generates a closure and sends it to the OpenFlow
controller as an NXT_CLOSURE asynchronous message (which must be
enabled through an extension to OFPT_SET_ASYNC or NXT_SET_ASYNC2).
The controller processes the closure, possibly modifying some of its
publicly accessible data (for now, the packet and its metadata), and
sends it back to the switch with an NXT_RESUME request, which
causes flow table traversal to continue.  In principle, a single
packet can be paused and resumed multiple times.

Another way to look at it is:

    - NXAST_PAUSE is an extension of the existing OFPAT_CONTROLLER
      action.  It sends the packet to the controller, with full
      pipeline context (some of which is switch implementation
      dependent, and may thus vary from switch to switch).

    - NXT_CLOSURE is an extension of OFPT_PACKET_IN, allowing for
      implementation dependent metadata.

    - NXT_RESUME is an extension of OFPT_PACKET_OUT, with the
      semantics that the pipeline processing is continued with the
      original translation context from where it was left at the time
      of NXAST_PAUSE.

Signed-off-by: Ben Pfaff <blp@ovn.org>