[RFC] raft: New module to implement Raft consensus algorithm.
The implementation is somewhat skeletal.

For testing, this adds the model checking and deque libraries from GNU
PSPP.  Their license is GPLv3+, but that should not be a problem because
they are not installed as part of Open vSwitch; they are for testing only.
I am the original author of these libraries.

Signed-off-by: Ben Pfaff <blp@nicira.com>