ovn-nbctl: Add "sync" command to wait for previous changes to take effect.
It's slow to add --wait to every ovn-nbctl command; only the last command
needs it.  But it's sometimes inconvenient to add it to the last command
if it's in a loop, etc.  This makes it possible to separately wait for
the OVN southbound or hypervisors to catch up to the northbound.

Also, modify the tests to replace "sleep" invocations that were waiting
for hypervisors to catch up by --wait=hv or by invocations of "ovn-nbctl
--wait=hv sync".

Signed-off-by: Ben Pfaff <blp@ovn.org>