ovn: support for service function chaining
This patchset is the first round at having Service Function Chaining
functionality through OVN. The implementation is done entirely
on the northbound side of OVN. It is a bump on the wire implementation,
so no attempt is being made in keeping state while packets visit each
hop of the chain. ACLs are used as the classifiers, with the augmentation
of action SFC, as well as option column.

The current implementation of traffic redirection to the service chain
is implemented by adding an additional action 'sfc' to the ACL stage. This
overloads the ACL stage and this might not be the best long term approach.
Guidence on whether this is "good enough" for now would be appreciated.

How to leverage load balancing is also an open issue. The current LB solution
in OVN is L3 based. Suggestions on how to implement LB at L2 for SFC would
also be appreciated.

This is not yet ready to be merged, as it lacks unit tests and a rigorous
code review. Nevertheless, it works fine if you take into account a
number of limitations that include:

* missing load balancer integration;
* no ipv6 support;
* chain spanning logical switches (not supported);
* bidirectional chains (not implemented);
* no test cases.
* other suggestions?

This is the code that was used for SFC demo and talk at OVSCon 2016.

Changes:

 * ovn-northd.xml: Added documentation for SFC ACL Action
 * ovn-northd.c: Added new stage for SFC and modified ACL stage to include sfc action
 * ovn-architecture.xml Included architecture of SFC in documentation
 * ovn-nb.ovsschema: Extended schema to include port-chain, port-pair-groups, port-pairs
    and added ACL SFC action
 * ovn-nb.xml: Added documentation for extensions to  ovn-nbctl for port-chain,
    port-pair-groups, port-pairs and ACL SFC action
 * ovn-nbctl.c: Added code to extend ovn-nbctl for port-chain,
    port-pair-groups, port-pairs and ACL SFC action

Current Limitations

This is not yet ready to be merged, as it lacks unit tests and a rigorous
code review. Nevertheless, it works fine if you take into account a
few limitations:

* missing load balancer integration.
* no ipv6 support.
* chain spanning logical switches (not supported).
* bidirectional chains (not implemented).
* no test cases.
* documentation needs rework as there have been several changes as the code has progressed.

Before I work on the limitations and start adding test cases I would like to make sure I am on the right track to get this approved for submission. Once it is approved for OVN/OVS I can add it to Openstack and also planning on using it for container service chaining.

Questions:

* Is the basic approach aligned with the direction of OVN/OVS.
* Is the approach to overloading the current ACL stage best approach for added a classifier action.
* How to approach load balancing - current approach in OVN is L# based any ideas on how to LB L2 virtual services.

Co-authored-by: Flavio Fernandes <flavio at flaviof.com>
Reported at: https://mail.openvswitch.org/pipermail/ovs-discuss/2016-March/040381.html
Reported at: https://mail.openvswitch.org/pipermail/ovs-discuss/2016-May/041359.html

Signed-off-by: John McDowall <jmcdowall at paloaltonetworks.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>