ovsdb: Remove and stop using dot2pic script.
Back in 2010, OVS stopped using graphviz's built-in "pic" output format
because the recent versions of Graphviz didn't seem to support it any
longer.  At that point, we switched to a home-grown script to convert
graphviz output to pic format.

However, the now-current versions of it do support "pic" again, and their
output is much nicer than dot2pic output, mostly because dot2pic output
doesn't give any font size hints so that text tends to escape from its box.
This commit removes the dot2pic translator and starts using the built-in
output format again.  It also makes the configure test for graphviz test
that it supports "pic" output format, so that if someone is using a "dot"
binary that lacks it, they simply won't get the diagrams, as if they
didn't have Graphviz at all.

Signed-off-by: Ben Pfaff <blp@ovn.org>