ovsdb: Drop distinction between monitors and replicas.
Until now, OVSDB distinguished between "monitors", which are associated
with OVSDB JSON-RPC client sessions and allow clients to find out about
database changes, and "replicas", which are associated with databases
and also find out about database changes and act on them in some way.
The rationale was that there could be more than one kind of replica.
However, in practice, replicas were only used by monitors, except for
updating the on-disk representation of the database (and that's now
broken out into a separate concept), so this commit merges replicas
into monitors to make the code easier to follow.

Signed-off-by: Ben Pfaff <blp@ovn.org>