ovsdb-server: Distinguish logs from other replicas.
Until now, ovsdb-server has internally chained a list of replicas from
each database.  Whenever a transaction commits, it is passed to each
of the replicas.  The first replica, which until now is always the disk
file that stores the database, is special because it is the only replica
that can report an error and thereby abort the transaction.  This commit
breaks that first replica out as a separate kind of entity that is not
on the list of replicas.  In later commits when we start adding support
for clustering, there will only be more and more special cases for the
"first replica", so it makes sense to distinguish it this way.

Signed-off-by: Ben Pfaff <blp@ovn.org>